+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Not released
 - Histogram tooltip formatter receiving dataIndex and ticks [#220](https://github.com/CartoDB/carto-react/pull/220)
+- Histogram yAxis max value should always be shown [#221](https://github.com/CartoDB/carto-react/pull/221)
 
 ## 1.1.0-beta.3 (2021-10-27)